Original change: https://android-review.googlesource.com/c/platform/packages/modules/Bluetooth/+/2908981

+++ android/app/src/com/android/bluetooth/mcp/MediaControlGattService.java
@@ -47,6 +47,7 @@ import com.android.bluetooth.BluetoothEventLogger;
 import com.android.bluetooth.Utils;
 import com.android.bluetooth.a2dp.A2dpService;
 import com.android.bluetooth.btservice.AdapterService;
+import com.android.bluetooth.flags.Flags;
 import com.android.bluetooth.hearingaid.HearingAidService;
 import com.android.bluetooth.le_audio.LeAudioService;
 import com.android.internal.annotations.VisibleForTesting;
@@ -1222,17 +1223,19 @@ public class MediaControlGattService implements MediaControlGattServiceInterface
                         + " request up");
 
         // TODO: Activate/deactivate devices with ActiveDeviceManager
-        if (req.getOpcode() == Request.Opcodes.PLAY) {
+        if (mLeAudioService == null) {
+            mLeAudioService = LeAudioService.getLeAudioService();
+        }
+        if (!isBroadcastActive() && req.getOpcode() == Request.Opcodes.PLAY) {
             if (mAdapterService.getActiveDevices(BluetoothProfile.A2DP).size() > 0) {
                 A2dpService.getA2dpService().removeActiveDevice(false);
             }
             if (mAdapterService.getActiveDevices(BluetoothProfile.HEARING_AID).size() > 0) {
                 HearingAidService.getHearingAidService().removeActiveDevice(false);
             }
-            if (mLeAudioService == null) {
-                mLeAudioService = LeAudioService.getLeAudioService();
+            if (mLeAudioService != null) {
+                mLeAudioService.setActiveDevice(device);
             }
-            mLeAudioService.setActiveDevice(device);
         }
         mCallbacks.onMediaControlRequest(req);
 
@@ -2051,6 +2054,20 @@ public class MediaControlGattService implements MediaControlGattServiceInterface
                     + ((mFeatures & featureBit) != 0));
         }
         return (mFeatures & featureBit) != 0;
+    }
+
+    /**
+     * Checks if le audio broadcasting is ON
+     *
+     * @return {@code true} if is broadcasting audio, {@code false} otherwise
+     */
+    private boolean isBroadcastActive() {
+        if (!Flags.leaudioBroadcastFeatureSupport()) {
+            // disable this if feature flag is false
+            return false;
+        }
+
+        return mLeAudioService != null && mLeAudioService.isBroadcastActive();
     }
 
     @VisibleForTesting
